Add ARIA-Live to Showing x of y on product filter
Issue: #148

* Had to remove the Handlebars approach and replace with a simple string
and some code re-org. After a lot of investigation, I found that
ARIA-Live doesn't work well if the replaced contents include HTML
elements. Text-only swaps are ideal, otherwise there's unwanted
side-effects like the read-order changing or the SR will only read the
content that's changed (despite different settings for `ARIA-atomic`).

* Because of the above issue, we couldn't use spans anymore to hide the
'to' visually and hide the dash from the SR. So, I had to change:
`1 - 12 of 40` to:
`1 to 12 of 40`

* Removed now deprecated showing-count.hbs