Pagination needs 1 …5…10 support #124
Improve the functionality of the pagination and limit the number of page number can be showned in the pagination added the ... for the hidden pages. Plus also refactor the code to generate the data now data generate code will generate the products_data.json file as well with product.csv file.